packer: linux: Support building for x86_64 and arm64
Introduce some user variables to select between x86_64 and arm64.  Since
the template support in packer is limited we make both arch and
instance_type a variable (even though arch effectively implies instance
type).

We update the Makefile to support ARCH=x86_64 or ARCH=arm64 and it will
set the instance_type for us.

We also update the ami_name to include the arch type so its a little
easier to tell the difference in AMIs.

Signed-off-by: Kumar Gala <kumar.gala@linaro.org>